removed elementView from model attributes
Was storing a reference to the element view on the model for improved DOM performance. I moved this reference from the attributes to the model envelope itself. This was mainly to prevent the reference from triggering a change event and trying to send the view to the parent window via the postMessage api.